server/requirements.txt
# Note: Python 3.11 or 3.12 recommended for best compatibility
# Python 3.14 has compatibility issues with pydantic-core compilation
fastapi==0.115.0
uvicorn[standard]==0.32.0
pyyaml==6.0.2
simpleeval==0.9.13
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
httpx==0.27.0

